+++ packages/plugin/src/classes/helpers/classes.js
@@ -8,6 +8,7 @@ import * as ast from "../../utils/ast";
 
 import { getJsDocClassInfo, getTags } from "./jsdoc";
 import { getDecoratorClassInfo } from "./decorators";
+import { getImportDeclaration } from "./imports";
 
 /**
  * Converts an ES6 class to a UI5 extend.
@@ -19,12 +20,14 @@ export function convertClassToUI5Extend(
   node,
   classInfo,
   extraStaticProps,
+  importDeclarationPaths,
   opts
 ) {
   if (!(t.isClassDeclaration(node) || t.isClassExpression(node))) {
     return node;
   }
 
+  const CONTROLLER_EXTENSION_TAG = "transformControllerExtension";
   const staticMembers = [];
 
   const classNameIdentifier = node.id;
@@ -65,17 +68,57 @@ export function convertClassToUI5Extend(
     }
   }
 
-  for (const member of node.body.body) {
+  for (const memberPath of path.get("body.body")) {
+    const member = memberPath.node;
     const memberName = member.key.name;
 
     if (t.isClassMethod(member)) {
+      const isConstructor = member.kind === "constructor";
+      const membersToAssign = [];
+      const params = isConstructor
+        ? member.params?.map((param) => {
+            // handling of parameter properties for constructors (TypeScript):
+            // https://www.typescriptlang.org/docs/handbook/2/classes.html#parameter-properties
+            //   -> extracting the real parameters and store the members to assign
+            if (param.type === "TSParameterProperty") {
+              membersToAssign.push(param.parameter);
+              return param.parameter;
+            }
+            return param;
+          })
+        : member.params;
       const func = t.functionExpression(
         member.key,
-        member.params,
+        params,
         member.body,
         member.generator,
         member.async
       );
+      if (isConstructor && membersToAssign.length > 0) {
+        // handling of parameter properties for constructors (TypeScript):
+        //   -> assigning parameter properties as members to the instance
+        const newMembers = membersToAssign.map((member) =>
+          buildMemberAssignmentStatement(t.thisExpression(), {
+            key: member,
+            computed: false,
+            value: member,
+          })
+        );
+        const superIndex = member.body.body.findIndex(
+          (node) =>
+            ast.isSuperCallExpression(node.expression) ||
+            ast.isSuperPrototypeCallOf(
+              node.expression,
+              superClassName,
+              "constructor"
+            )
+        );
+        member.body.body.splice(
+          superIndex === -1 ? member.body.body.length : superIndex + 1,
+          0,
+          ...newMembers
+        );
+      }
       if (member.static) {
         staticMembers.push(
           buildMemberAssignmentStatement(classNameIdentifier, {
@@ -114,7 +157,60 @@ export function convertClassToUI5Extend(
         }
       }
     } else if (t.isClassProperty(member)) {
-      if (!member.value) continue; // un-initialized static class prop (typescript)
+      // For class properties annotated to represent controller extensions, replace the pure declaration with an assignment (that's what the runtime expects)
+      // and keep them at the initialization object as properties (don't move into constructor).
+      if (
+        member.leadingComments?.some((comment) => {
+          return comment.value.includes("@" + CONTROLLER_EXTENSION_TAG);
+        }) ||
+        member.decorators?.some((decorator) => {
+          return decorator.expression?.name === CONTROLLER_EXTENSION_TAG;
+        })
+      ) {
+        const typeAnnotation = member.typeAnnotation?.typeAnnotation;
+        // double-check that it is a valid node for a controller extension
+        if (
+          t.isTSTypeReference(typeAnnotation) ||
+          t.isTSQualifiedName(typeAnnotation)
+        ) {
+          const typeName = getTypeName(typeAnnotation);
+
+          // 1. transform the property from being typed as instance and un-initialized to a property where the controller extension *class* is assigned as value
+          const valueIdentifier = t.identifier(typeName);
+          member.value = valueIdentifier;
+          member.typeAnnotation = null;
+          extendProps.unshift(buildObjectProperty(member)); // add it to the properties of the extend() config object
+
+          // 2. add a binding reference to the value, so in case the TS transpiler runs later it recognizes that the import is still needed
+          const typeNameFirstPart = typeName.split(".")[0]; // e.g. when "myExtension: someBundle.MyExtension"
+          if (memberPath.scope.hasBinding(typeNameFirstPart)) {
+            const binding = path.scope.getBinding(typeNameFirstPart);
+            binding.referencePaths.push(memberPath.get("value"));
+          }
+
+          // 3. restore the import in case it was run already and removed the import
+          const neededImportDeclaration = getImportDeclaration(
+            memberPath.hub.file.opts.filename,
+            typeName
+          );
+          if (
+            !importDeclarationPaths.some(
+              (path) => path.node === neededImportDeclaration
+            )
+          ) {
+            // TODO: import might be there but with the specifier removed; we can clone, but then other specifiers are duplicate
+            // if import is no longer there, re-add it
+            importDeclarationPaths[
+              importDeclarationPaths.length - 1
+            ].insertAfter(neededImportDeclaration);
+          }
+
+          // 4. prevent the member from also being added to the constructor (member does have a value now and initializer would be added below)
+          continue;
+        }
+      }
+
+      if (!member.value) continue; // remove all other un-initialized static class props (typescript)
 
       // Special handling for TypeScript limitation where metadata, renderer and overrides must be properties.
       if (["metadata", "renderer", "overrides"].includes(memberName)) {
@@ -205,7 +301,15 @@ export function convertClassToUI5Extend(
       // Copy all except the super call from the constructor to the bindMethod (i.e. onInit)
       bindMethod.body.body.unshift(
         ...constructor.body.body.filter(
-          (node) => !ast.isSuperCallExpression(node.expression)
+          (node) =>
+            !(
+              ast.isSuperCallExpression(node.expression) ||
+              ast.isSuperPrototypeCallOf(
+                node.expression,
+                superClassName,
+                "constructor"
+              )
+            )
         )
       );
     }
@@ -291,17 +395,55 @@ function getFileBaseNamespace(path, pluginOpts) {
   }
 }
 
-const buildObjectProperty = (member) =>
-  t.objectProperty(member.key, member.value, member.computed);
+const getQualifiedName = (node) => {
+  let { left, right } = node;
 
-const buildMemberAssignmentStatement = (objectIdentifier, member) =>
-  t.expressionStatement(
+  // if left is TSQualifiedName, recursive call to get full namespace
+  if (t.isTSQualifiedName(left)) {
+    left = getQualifiedName(left);
+  } else {
+    // if left is an Identifier
+    left = left.name;
+  }
+
+  return `${left}.${right.name}`;
+};
+
+export const getTypeName = (typeAnnotation) => {
+  if (t.isTSTypeReference(typeAnnotation)) {
+    // for TSTypeReference, typeName can be an Identifier or a TSQualifiedName
+    return (
+      typeAnnotation.typeName.name || getQualifiedName(typeAnnotation.typeName)
+    );
+  }
+  if (t.isTSQualifiedName(typeAnnotation)) {
+    // for TSQualifiedName
+    return getQualifiedName(typeAnnotation);
+  }
+  return null;
+};
+
+const buildObjectProperty = (member) => {
+  const newObjectProperty = t.objectProperty(
+    member.key,
+    member.value,
+    member.computed
+  );
+  newObjectProperty.leadingComments = member.leadingComments;
+  return newObjectProperty;
+};
+
+const buildMemberAssignmentStatement = (objectIdentifier, member) => {
+  const newMember = t.expressionStatement(
     t.assignmentExpression(
       "=",
       t.memberExpression(objectIdentifier, member.key, member.computed),
       member.value
     )
   );
+  newMember.leadingComments = member.leadingComments;
+  return newMember;
+};
 
 const buildThisMemberAssignmentStatement = buildMemberAssignmentStatement.bind(
   null,