+++ packages/plugin/src/classes/visitor.js
@@ -6,6 +6,9 @@ import * as classes from "./helpers/classes";
 const CONSTRUCTOR = "constructor";
 
 export const ClassTransformVisitor = {
+  ImportDeclaration(path) {
+    this.importDeclarationPaths.push(path);
+  },
   ImportDefaultSpecifier(path) {
     this.importNames.push(path.node.local.name);
   },
@@ -17,9 +20,10 @@ export const ClassTransformVisitor = {
    */
   Class: {
     enter(path, { file, opts = {} }) {
-      const { node, parent, parentPath } = path;
+      const { node } = path;
+      const className = node?.id?.name;
 
-      if (opts.neverConvertClass) {
+      if (!className || opts.neverConvertClass) {
         return;
       }
       if (!doesClassExtendFromImport(node, [...this.importNames])) {
@@ -31,25 +35,46 @@ export const ClassTransformVisitor = {
       // and therefore may need to be transformed to .extend() syntax.
       const classInfo = classes.getClassInfo(path, node, path.parent, opts);
 
-      if (!shouldConvertClass(file, node, opts, classInfo)) {
+      // filter found decorators (to be not handled by other plugins)
+      node.decorators = node.decorators?.filter((d) => {
+        const exp = d.expression;
+        const name = exp.name || exp.callee?.name;
+        return classInfo[name] === undefined;
+      });
+
+      if (shouldConvertClass(file, node, opts, classInfo)) {
+        // Save super class name for converting super calls
+        this.superClassName = classInfo.superClassName;
+        // store the classinfo
+        if (className) {
+          this.classInfo = this.classInfo || {};
+          this.classInfo[className] = classInfo;
+        }
+      }
+    },
+    exit(path, { opts = {} }) {
+      const { node, parent, parentPath } = path;
+      const className = node?.id?.name;
+
+      // Only if classinfo has been found we process this file
+      const classInfo = this.classInfo?.[className];
+      if (!classInfo || !node.superClass) {
         return;
       }
 
-      // Save super class name for converting super calls
-      this.superClassName = classInfo.superClassName;
-
       // Find the Block scoped parent (Program or Function body) and search for assigned properties within that (eg. MyClass.X = "X").
-      const { name: className } = node.id;
       const blockParent = path.findParent((path) => path.isBlock()).node;
       const staticProps = ast.groupPropertiesByName(
         ast.getOtherPropertiesOfIdentifier(blockParent, className)
       );
+
       // TODO: flag metadata and renderer for removal if applicable
       const ui5ExtendClass = classes.convertClassToUI5Extend(
         path,
         node,
         classInfo,
         staticProps,
+        this.importDeclarationPaths,
         opts
       );
 
@@ -77,8 +102,7 @@ export const ClassTransformVisitor = {
         }
         parentPath.replaceWithMultiple(ui5ExtendClass);
       }
-    },
-    exit() {
+
       this.superClassName = null;
     },
   },
@@ -138,7 +162,7 @@ function isSuperApply(callee) {
 function getRequiredParamsOfSAPUIDefine(path, node) {
   const defineArgs = node.arguments;
   const callbackNode = defineArgs.find((argNode) => t.isFunction(argNode));
-  return callbackNode.params; // Identifier
+  return callbackNode?.params || []; // Identifier
 }
 
 /**