+++ packages/preset/CHANGELOG.md
@@ -3,6 +3,113 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.4.2](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.4.1...v7.4.2) (2024-06-25)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.4.1](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.4.0...v7.4.1) (2024-06-02)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+# [7.4.0](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.3.1...v7.4.0) (2024-05-31)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.3.1](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.3.0...v7.3.1) (2024-05-12)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+# [7.3.0](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.6...v7.3.0) (2024-01-18)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.6](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.5...v7.2.6) (2023-11-23)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.5](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.4...v7.2.5) (2023-08-28)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.4](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.3...v7.2.4) (2023-07-20)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.3](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.2...v7.2.3) (2023-07-18)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.2](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.1...v7.2.2) (2023-07-18)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+## [7.2.1](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.2.0...v7.2.1) (2023-06-13)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
+# [7.2.0](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.1.5...v7.2.0) (2023-05-30)
+
+
+### Features
+
+* split class conversion to enable other Babel plugin conversion ([#100](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/issues/100)) ([4ba096b](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/commit/4ba096b1a24d807cda2fd2f57425f3ab4b91a31b)), closes [#23](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/issues/23) [#25](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/issues/25)
+
+
+
+
+
+## [7.1.5](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.1.4...v7.1.5) (2023-05-30)
+
+**Note:** Version bump only for package babel-preset-transform-ui5
+
+
+
+
+
 ## [7.1.4](https://github.com/ui5-community/babel-plugin-transform-modules-ui5/compare/v7.1.3...v7.1.4) (2023-04-11)
 
 **Note:** Version bump only for package babel-preset-transform-ui5